Udpated v8.2.1 : New files Required for proper Functionality
This update resolves several issues related to the database conversion process. The main problems stemmed from syntax errors that MySQL could handle but Microsoft SQL could not.

This version has been confirmed to be fully functional for:

LDAP Functionality

API Connections

Searching

GUI Navigation and changes

The database backup feature is currently not working. There is no backup script for SQL, and one will need to be developed. I am using an alternative backup method, so if you're interested in creating a script, you can take a shot at it.